refactor(zapier): improve code structure + use bot logs
